requirements.txt
asgiref==3.2.10
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.5
confusable-homoglyphs==3.2.0
Django==3.1.1
django-bootstrap-v5==1.0.5
django-registration==3.2
django-registration-redux==2.9
djangorestframework==3.12.4
idna==3.2
importlib-metadata==3.10.1
install==1.3.4
Pillow==8.3.2
psycopg2==2.7.5
python-dateutil==2.8.2
pytz==2021.1
pyuploadcare==2.7.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
